Add propert <repositories> sections to all the pom.xml files
In order to be able to import individual projects (as opposed to the
complete Fiji project including all sub projects), the pom.xml files
need to list the imagej.net Maven repositories so that the parent POM
can be found.

While at it, add the JBoss Maven repository to the pom.xml files of the
FS_Align_TrakEM2, Reconstruct_Reader, blockmatching_ and
register_virtual_slices projects so that JAI is found, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>